Added user authentication & user checking for CRUD functions

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use Illuminate\Http\Request;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -51,8 +61,9 @@ class PostsController extends Controller
 
         // Create Post
         $post = new Post;
-        $post->title = $request->input('title');
-        $post->body = $request->input('body');
+        $post->title = $request->input('title'); // adds title to post tuple
+        $post->body = $request->input('body'); // adds the body
+        $post->user_id = auth()->user()->id; // adds id of currently logged in user
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -79,6 +90,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        
+        // Checking for correct user
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Access');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -98,6 +115,12 @@ class PostsController extends Controller
 
         // Create Post
         $post = Post::find($id);
+
+        // Checking for correct user
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Access');
+        }
+
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->save();
@@ -114,6 +137,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Checking for correct user
+        if(auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Access');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
